+++ java-r5rcore/src/org/ipea/r5r/R5RCore.java
@@ -15,8 +15,6 @@ import org.ipea.r5r.Fares.RuleBasedInRoutingFareCalculator;
 import org.ipea.r5r.Modifications.R5RFileStorage;
 import org.ipea.r5r.Network.NetworkBuilder;
 import org.ipea.r5r.Process.*;
-import org.ipea.r5r.Utils.SpeedSetter;
-import org.ipea.r5r.Utils.SpeedSetter.SpeedSetterMode;
 import org.ipea.r5r.Utils.Utils;
 import org.slf4j.LoggerFactory;
 
@@ -701,5 +699,10 @@ public class R5RCore {
 
     public void abort() {
         LOG.error("process aborted");
+    }
+
+    /** Used to ensure user has passed a jobjRef to an R5R core in R. */
+    public String identify(){
+        return "I am an R5R core!";
     }
 }