+++ r-package/man/roxygen/templates/r5r_core.R
@@ -1,2 +1,2 @@
-#' @param r5r_core An object to connect with the R5 routing engine, created with
-#' [setup_r5()].
+#' @param r5r_core The `r5r_core` argument is deprecated as of r5r v2.3.0.
+#'        Please use the `r5r_network` argument instead.