+++ r-package/tests/testthat/test-detailed_itineraries.R
@@ -5,7 +5,7 @@ context("Detailed itineraries function")
 
 testthat::skip_on_cran()
 
-default_tester <- function(r5r_core,
+default_tester <- function(r5r_network,
                            origins = points[1:2, ],
                            destinations = points[2:1, ],
                            mode = c("WALK", "TRANSIT"),
@@ -24,7 +24,7 @@ default_tester <- function(r5r_core,
                            osm_link_ids = FALSE) {
 
  results <- detailed_itineraries(
-   r5r_core,
+   r5r_network,
    origins = origins,
    destinations = destinations,
    mode = mode,
@@ -52,8 +52,8 @@ default_tester <- function(r5r_core,
 
 test_that("detailed_itineraries adequately raises errors", {
 
-  # error related to using object with wrong type as r5r_core
-  expect_error(default_tester("r5r_core"))
+  # error related to using object with wrong type as r5r_network
+  expect_error(default_tester("r5r_network"))
 
   # error related to using wrong origins/destinations object type
   multipoint_origins      <- sf::st_cast(sf::st_as_sf(points[1:2,], coords = c("lon", "lat")), "MULTIPOINT")
@@ -61,12 +61,12 @@ test_that("detailed_itineraries adequately raises errors", {
   list_origins      <- list(id = c("1", "2"), lat = c(-30.02756, -30.02329), long = c(-51.22781, -51.21886))
   list_destinations <- list(id = c("2", "1"), lat = c(-30.02329, -30.02756), long = c(-51.21886, -51.22781))
 
-  expect_error(default_tester(r5r_core, origins = multipoint_origins))
-  expect_error(default_tester(r5r_core, destinations = multipoint_destinations))
-  expect_error(default_tester(r5r_core, origins = list_origins))
-  expect_error(default_tester(r5r_core, destinations = list_destinations))
-  expect_error(default_tester(r5r_core, origins = "origins"))
-  expect_error(default_tester(r5r_core, destinations = "destinations"))
+  expect_error(default_tester(r5r_network, origins = multipoint_origins))
+  expect_error(default_tester(r5r_network, destinations = multipoint_destinations))
+  expect_error(default_tester(r5r_network, origins = list_origins))
+  expect_error(default_tester(r5r_network, destinations = list_destinations))
+  expect_error(default_tester(r5r_network, origins = "origins"))
+  expect_error(default_tester(r5r_network, destinations = "destinations"))
 
   # error related to using wrong origins/destinations column types
   origins <- points[1:2, ]
@@ -77,67 +77,67 @@ test_that("detailed_itineraries adequately raises errors", {
   destinations_char_lat   <- data.frame(id = destinations$id, lat = as.character(destinations$lat), lon = destinations$lon)
   destinations_char_lon   <- data.frame(id = destinations$id, lat = destinations$lat, lon = as.character(destinations$lon))
 
-  expect_error(default_tester(r5r_core, origins = origins_char_lat))
-  expect_error(default_tester(r5r_core, origins = origins_char_lon))
-  expect_error(default_tester(r5r_core, destinations = destinations_char_lat))
-  expect_error(default_tester(r5r_core, destinations = destinations_char_lon))
+  expect_error(default_tester(r5r_network, origins = origins_char_lat))
+  expect_error(default_tester(r5r_network, origins = origins_char_lon))
+  expect_error(default_tester(r5r_network, destinations = destinations_char_lat))
+  expect_error(default_tester(r5r_network, destinations = destinations_char_lon))
 
   # error related to 'origins' and 'destinations' with distinct number of rows
-  expect_error(default_tester(r5r_core, origins = points[1:3,], destinations = points[2:1,]))
-  expect_error(default_tester(r5r_core, origins = points[2:1,], destinations = points[1:3,]))
+  expect_error(default_tester(r5r_network, origins = points[1:3,], destinations = points[2:1,]))
+  expect_error(default_tester(r5r_network, origins = points[2:1,], destinations = points[1:3,]))
 
   # error related to nonexistent mode
-  expect_error(default_tester(r5r_core, mode = "all"))
+  expect_error(default_tester(r5r_network, mode = "all"))
 
   # errors related to date formatting
   numeric_datetime <- as.numeric(as.POSIXct("13-05-2019 14:00:00", format = "%d-%m-%Y %H:%M:%S"))
 
-  expect_error(default_tester(r5r_core, departure_datetime = "13-05-2019 14:00:00"))
-  expect_error(default_tester(r5r_core, numeric_datetime))
+  expect_error(default_tester(r5r_network, departure_datetime = "13-05-2019 14:00:00"))
+  expect_error(default_tester(r5r_network, numeric_datetime))
 
   # errors related to max_walk_time
-  expect_error(default_tester(r5r_core, max_walk_time = "1"))
-  expect_error(default_tester(r5r_core, max_walk_time = NULL))
+  expect_error(default_tester(r5r_network, max_walk_time = "1"))
+  expect_error(default_tester(r5r_network, max_walk_time = NULL))
 
   # errors related to max_bike_time
-  expect_error(default_tester(r5r_core, max_bike_time = "1"))
-  expect_error(default_tester(r5r_core, max_bike_time = NULL))
+  expect_error(default_tester(r5r_network, max_bike_time = "1"))
+  expect_error(default_tester(r5r_network, max_bike_time = NULL))
 
     # error/warning related to max_street_time
-  expect_error(default_tester(r5r_core, max_trip_duration = "120"))
+  expect_error(default_tester(r5r_network, max_trip_duration = "120"))
 
   # error related to non-numeric walk_speed
-  expect_error(default_tester(r5r_core, walk_speed = "3.6"))
+  expect_error(default_tester(r5r_network, walk_speed = "3.6"))
 
   # error related to non-numeric bike_speed
-  expect_error(default_tester(r5r_core, bike_speed = "12"))
+  expect_error(default_tester(r5r_network, bike_speed = "12"))
 
   # error related to non-logical shortest_path
-  expect_error(default_tester(r5r_core, shortest_path = "TRUE"))
-  expect_error(default_tester(r5r_core, shortest_path = 1))
-  expect_error(default_tester(r5r_core, shortest_path = NULL))
+  expect_error(default_tester(r5r_network, shortest_path = "TRUE"))
+  expect_error(default_tester(r5r_network, shortest_path = 1))
+  expect_error(default_tester(r5r_network, shortest_path = NULL))
 
   # error related to non-logical verbose
-  expect_error(default_tester(r5r_core, verbose = "TRUE"))
-  expect_error(default_tester(r5r_core, verbose = 1))
-  expect_error(default_tester(r5r_core, verbose = NULL))
+  expect_error(default_tester(r5r_network, verbose = "TRUE"))
+  expect_error(default_tester(r5r_network, verbose = 1))
+  expect_error(default_tester(r5r_network, verbose = NULL))
 
   # error related to non-logical drop_geometry
-  expect_error(default_tester(r5r_core, drop_geometry = "TRUE"))
-  expect_error(default_tester(r5r_core, drop_geometry = 1))
+  expect_error(default_tester(r5r_network, drop_geometry = "TRUE"))
+  expect_error(default_tester(r5r_network, drop_geometry = 1))
 
   # error related to non-logical osm_link_ids
-  expect_error(default_tester(r5r_core, osm_link_ids = "TRUE"))
-  expect_error(default_tester(r5r_core, osm_link_ids = 1))
-  expect_error(default_tester(r5r_core, osm_link_ids = TRUE, drop_geometry=TRUE))
+  expect_error(default_tester(r5r_network, osm_link_ids = "TRUE"))
+  expect_error(default_tester(r5r_network, osm_link_ids = 1))
+  expect_error(default_tester(r5r_network, osm_link_ids = TRUE, drop_geometry=TRUE))
 
   })
 
 test_that("detailed_itineraries adequately raises warnings and messages", {
 
   # message related to expanding origins/destinations dataframe
-  expect_message(default_tester(r5r_core, origins = points[1, ]))
-  expect_message(default_tester(r5r_core, destinations = points[1, ]))
+  expect_message(default_tester(r5r_network, origins = points[1, ]))
+  expect_message(default_tester(r5r_network, destinations = points[1, ]))
 
   # error/warning related to using wrong origins/destinations column types
   origins <- points[1:2, ]
@@ -146,8 +146,8 @@ test_that("detailed_itineraries adequately raises warnings and messages", {
   origins_numeric_id <- data.frame(id = 1:2, lat = origins$lat, lon = origins$lon)
   destinations_numeric_id <- data.frame(id = 1:2, lat = destinations$lat, lon = destinations$lon)
 
-  expect_warning(default_tester(r5r_core, origins = origins_numeric_id))
-  expect_warning(default_tester(r5r_core, destinations = destinations_numeric_id))
+  expect_warning(default_tester(r5r_network, origins = origins_numeric_id))
+  expect_warning(default_tester(r5r_network, destinations = destinations_numeric_id))
 
 
 })
@@ -175,8 +175,8 @@ test_that("detailed_itineraries output is correct", {
     crs = 4326
   )
 
-  result_df_input <- default_tester(r5r_core)
-  result_sf_input <- default_tester(r5r_core, origins_sf, destinations_sf)
+  result_df_input <- default_tester(r5r_network)
+  result_sf_input <- default_tester(r5r_network, origins_sf, destinations_sf)
 
   expect_true(is(result_df_input, "sf"))
   expect_true(is(result_df_input, "data.table"))
@@ -186,8 +186,8 @@ test_that("detailed_itineraries output is correct", {
   # expect results to be of class 'data.table', but not 'sf', independently of
   # the class of 'origins'/'destinations', when drop_geometry = TRUE
 
-  result_df_input <- default_tester(r5r_core, drop_geometry = TRUE)
-  result_sf_input <- default_tester(r5r_core, origins_sf, destinations_sf, drop_geometry = TRUE)
+  result_df_input <- default_tester(r5r_network, drop_geometry = TRUE)
+  result_sf_input <- default_tester(r5r_network, origins_sf, destinations_sf, drop_geometry = TRUE)
 
   expect_false(is(result_df_input, "sf"))
   expect_true(is(result_df_input, "data.table"))
@@ -195,7 +195,7 @@ test_that("detailed_itineraries output is correct", {
   expect_true(is(result_sf_input, "data.table"))
 
   # bring osm ids
-  result_df_input_ids <- default_tester(r5r_core,
+  result_df_input_ids <- default_tester(r5r_network,
                                         osm_link_ids = TRUE,
                                         drop_geometry = FALSE)
 
@@ -235,10 +235,10 @@ test_that("detailed_itineraries output is correct", {
   origins      <- points[1,]
   destinations <- points[3,]
 
-  df <- default_tester(r5r_core, origins = origins, destinations = destinations, mode = "WALK", walk_speed = 4)
+  df <- default_tester(r5r_network, origins = origins, destinations = destinations, mode = "WALK", walk_speed = 4)
   duration_lower_speed <- data.table::setDT(df)$segment_duration
 
-  df <- default_tester(r5r_core, origins = origins, destinations = destinations, mode = "WALK", walk_speed = 5)
+  df <- default_tester(r5r_network, origins = origins, destinations = destinations, mode = "WALK", walk_speed = 5)
   duration_higher_speed <- data.table::setDT(df)$segment_duration
 
   expect_true(duration_higher_speed < duration_lower_speed)
@@ -246,11 +246,11 @@ test_that("detailed_itineraries output is correct", {
   # expect bike segments to be shorter when setting higher cycling speeds
   # ps: same as with walk_speeds
 
-  df <- default_tester(r5r_core, origins = origins, destinations = destinations,
+  df <- default_tester(r5r_network, origins = origins, destinations = destinations,
                        mode = "BICYCLE", bike_speed = 12)
   duration_lower_speed <- data.table::setDT(df)$segment_duration
 
-  df <- default_tester(r5r_core, origins = origins, destinations = destinations,
+  df <- default_tester(r5r_network, origins = origins, destinations = destinations,
                        mode = "BICYCLE", bike_speed = 13)
   duration_higher_speed <- data.table::setDT(df)$segment_duration
 
@@ -262,14 +262,14 @@ test_that("detailed_itineraries output is correct", {
 
   # expect each OD pair to have only option when shortest_path == TRUE
 
-  df <- default_tester(r5r_core, shortest_path = TRUE)
+  df <- default_tester(r5r_network, shortest_path = TRUE)
   max_n_options <- data.table::setDT(df)[, length(unique(option)), by = .(from_id, to_id)][, max(V1)]
 
   expect_true(max_n_options == 1)
 
   # # expect each OD pair to have (possibly) more than one option when shortest_path == FALSE
   #
-  # df <- default_tester(r5r_core, shortest_path = FALSE)
+  # df <- default_tester(r5r_network, shortest_path = FALSE)
   # max_n_options <- data.table::setDT(df)[, length(unique(option)), by = .(from_id, to_id)][, max(V1)]
   #
   # expect_true(max_n_options > 1)
@@ -280,7 +280,7 @@ test_that("detailed_itineraries output is correct", {
   origins <- points[1:15,]
   destinations <- points[15:1,]
 
-  df <- default_tester(r5r_core, origins, destinations,
+  df <- default_tester(r5r_network, origins, destinations,
                        max_trip_duration = max_trip_duration, shortest_path = FALSE)
 
   max_duration <- data.table::setDT(df)[, sum(segment_duration), by = .(from_id, to_id, option)][, max(V1)]
@@ -289,12 +289,12 @@ test_that("detailed_itineraries output is correct", {
 
   # expect an empty data.table as output when no routes are found between the pairs
 
-  df <- default_tester(r5r_core, origins = points[10,], destinations = points[12,],
+  df <- default_tester(r5r_network, origins = points[10,], destinations = points[12,],
                        mode = "WALK", max_trip_duration = 30L)
 
   expect_true(nrow(df) == 0)
 
-  df <- default_tester(r5r_core, origins = points[10:11,], destinations = points[12,],
+  df <- default_tester(r5r_network, origins = points[10:11,], destinations = points[12,],
                        mode = "WALK", max_trip_duration = 30L)
 
   expect_true(nrow(df) == 0)
@@ -304,7 +304,7 @@ test_that("detailed_itineraries output is correct", {
 
 test_that("using transit outside the gtfs dates throws an error", {
   expect_error(
-    tester(r5r_core,
+    tester(r5r_network,
            mode='transit',
            departure_datetime = as.POSIXct("13-05-2025 14:00:00",
                                            format = "%d-%m-%Y %H:%M:%S")