+++ r-package/tests/testthat/test-set_percentiles.R
@@ -3,7 +3,9 @@
 
 testthat::skip_on_cran()
 
-tester <- function(percentiles) set_percentiles(r5r_core, percentiles)
+tester <- function(percentiles) {
+  set_percentiles(r5r_network, percentiles)
+  }
 
 test_that("input is correct", {
   expect_error(tester("1"))
@@ -18,9 +20,9 @@ test_that("input is correct", {
 test_that("set_percentiles argument works in travel_time_matrix()", {
   basic_expr <- call(
     "travel_time_matrix",
-    r5r_core,
-    pois,
-    pois,
+    r5r_network,
+    origins=pois,
+    destinations=pois,
     mode = c("TRANSIT", "WALK"),
     departure_datetime = departure_datetime,
     draws_per_minute = 1,
@@ -66,9 +68,9 @@ test_that("set_percentiles argument works in travel_time_matrix()", {
 test_that("set_percentiles argument works in accessibility()", {
   basic_expr <- call(
     "accessibility",
-    r5r_core,
-    points[1:30],
-    points[1:30],
+    r5r_network,
+    origins=points[1:30],
+    destinations=points[1:30],
     mode = c("TRANSIT", "WALK"),
     departure_datetime = departure_datetime,
     opportunities_colnames = "population",
@@ -108,9 +110,9 @@ test_that("set_percentiles argument works in accessibility()", {
 test_that("set_percentiles argument works in pareto_frontier()", {
   basic_expr <- call(
     "pareto_frontier",
-    r5r_core,
-    points[1:5],
-    points[1:5],
+    r5r_network,
+    origins=points[1:5],
+    destinations=points[1:5],
     mode = c("TRANSIT", "WALK"),
     departure_datetime = departure_datetime,
     fare_structure = fare_structure,