+++ r-package/tests/testthat/test-travel_time_matrix.R
@@ -8,7 +8,7 @@ departure_datetime <- as.POSIXct(
   format = "%d-%m-%Y %H:%M:%S"
 )
 
-tester <- function(r5r_core = get("r5r_core", envir = parent.frame()),
+tester <- function(r5r_network = get("r5r_network", envir = parent.frame()),
                    origins = pois,
                    destinations = pois,
                    mode = "WALK",
@@ -31,7 +31,7 @@ tester <- function(r5r_core = get("r5r_core", envir = parent.frame()),
                    progress = FALSE,
                    output_dir = NULL) {
   travel_time_matrix(
-    r5r_core,
+    r5r_network,
     origins = origins,
     destinations = destinations,
     mode = mode,
@@ -90,7 +90,7 @@ test_that("errors due to incorrect input types - origins and destinations", {
 test_that("errors due to incorrect input types - other inputs", {
   # mode and mode_egress are tested in assign_mode() tests
 
-  expect_error(tester(unclass(r5r_core)))
+  expect_error(tester(unclass(r5r_network)))
 
   expect_error(tester(departure_datetime = unclass(departure_datetime)))
   expect_error(tester(departure_datetime = rep(departure_datetime, 2)))
@@ -289,7 +289,7 @@ test_that("returns ttm even if last call saved to dir", {
 
 test_that("using transit outside the gtfs dates throws an error", {
   expect_error(
-    tester(r5r_core,
+    tester(r5r_network,
            mode='transit',
            departure_datetime = as.POSIXct("13-05-2025 14:00:00",
                                            format = "%d-%m-%Y %H:%M:%S")