+++ r-package/NEWS.md
@@ -2,13 +2,16 @@
 
 **Major changes**
 
+- New function `build_network()` to replace `setup_r5()`, which is being deprecated.
 - New function `arrival_travel_time_matrix()` to calculate travel time matrix between origin destination pairs considering the time of arrival, instead of a depature time. Closes [#291](https://github.com/ipeaGIT/r5r/issues/291)
 - New function `modify_osm_carspeeds()` to build a routable network with modified OSM car speeds to account for different scenarios of traffic congestion and road closure. Closes [#289](https://github.com/ipeaGIT/r5r/issues/289)
 - We have now implemented a reverse search optimization for direct transport modes (walking and cycling) in the functions `travel_time_matrix()`, `expanded_travel_time_matrix()` and `arrival_travel_time_matrix()`. In pactice, this means that these functions are now much faster when there are multiple origins to few destinations. Closes [#450](https://github.com/ipeaGIT/r5r/issues/450)
 
 **Minor changes**
 
-- routing properties within little jar are reset to default after a routing execution [#453](https://github.com/ipeaGIT/r5r/pull/453)
+- The routable transport network build with `build_network()` and `setup_r5()` now have a their own class `"r5r_network"`, making the package more consistent and safer from errors. 
+- Routing properties within r5r jar (aka little jar) are reset to default after a routing execution [#453](https://github.com/ipeaGIT/r5r/pull/453)
+
 
 **Bug fixes**